# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BlogSmith",
    "description": "Agentic blogging service \u2014 a 9-stage LangGraph pipeline (Discovery \u2192 Research \u2192 Outline \u2192 Draft \u2192 Critique \u2192 human email gate \u2192 Finalize \u2192 Visuals \u2192 Distribute) that writes human-like, SEO-strong posts with Gemini image generation.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/account": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Get Account",
        "operationId": "get_account_account_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/account/keys": {
      "put": {
        "tags": [
          "account"
        ],
        "summary": "Update Keys",
        "operationId": "update_keys_account_keys_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderKeysIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/sites": {
      "get": {
        "tags": [
          "sites"
        ],
        "summary": "List Sites",
        "operationId": "list_sites_sites_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SiteOut"
                  },
                  "type": "array",
                  "title": "Response List Sites Sites Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "sites"
        ],
        "summary": "Create Site",
        "operationId": "create_site_sites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/sites/{site_id}": {
      "get": {
        "tags": [
          "sites"
        ],
        "summary": "Get Site",
        "operationId": "get_site_sites__site_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "sites"
        ],
        "summary": "Update Site",
        "operationId": "update_site_sites__site_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sites"
        ],
        "summary": "Delete Site",
        "operationId": "delete_site_sites__site_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sites/{site_id}/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "operationId": "create_run_sites__site_id__runs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_sites__site_id__runs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunOut"
                  },
                  "title": "Response List Runs Sites  Site Id  Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sites/{site_id}/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_sites__site_id__runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sites/{site_id}/runs/{run_id}/result": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Result",
        "operationId": "get_result_sites__site_id__runs__run_id__result_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountOut": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "plan": {
            "type": "string",
            "title": "Plan",
            "default": "free"
          },
          "keys": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Keys"
          }
        },
        "type": "object",
        "required": [
          "uid"
        ],
        "title": "AccountOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InternalLink": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords",
            "description": "Anchor-intent keywords for matching."
          }
        },
        "type": "object",
        "required": [
          "title",
          "url"
        ],
        "title": "InternalLink"
      },
      "ProviderKeysIn": {
        "properties": {
          "gemini_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gemini Key",
            "description": "Google Gemini API key (required to generate)."
          },
          "langsmith_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Langsmith Key",
            "description": "LangSmith key for tracing (optional)."
          },
          "serp_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serp Key",
            "description": "SERP provider key for discovery (optional)."
          },
          "sendgrid_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sendgrid Key",
            "description": "SendGrid key for sending approval emails (optional)."
          }
        },
        "type": "object",
        "title": "ProviderKeysIn",
        "description": "Keys a user pastes into their profile. All optional; only set ones update."
      },
      "RunCreate": {
        "properties": {
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic",
            "description": "Skip Discovery and write this exact topic."
          },
          "keyword": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyword",
            "description": "Primary target keyword (optional)."
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Freeform guidance for this single post."
          },
          "expert_insights": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expert Insights",
            "description": "Optional war story / real numbers / contrarian take to weave in (the E-E-A-T layer)."
          },
          "auto_approve": {
            "type": "boolean",
            "title": "Auto Approve",
            "description": "Skip the email gate (testing / trusted automation).",
            "default": false
          }
        },
        "type": "object",
        "title": "RunCreate"
      },
      "RunOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "keyword": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyword"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "stages": {
            "type": "object",
            "title": "Stages"
          }
        },
        "type": "object",
        "required": [
          "id",
          "site_id",
          "status"
        ],
        "title": "RunOut"
      },
...
      "SiteIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "brand_voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Voice"
          },
          "custom_prompts": {
            "$ref": "#/components/schemas/CustomPrompts"
          },
          "image_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Style",
            "description": "Visual style guidance for generated images."
          },
          "pillar_cluster_map": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Pillar Cluster Map",
            "description": "pillar -> cluster keywords."
          },
          "internal_links": {
            "items": {
              "$ref": "#/components/schemas/InternalLink"
            },
            "type": "array",
            "title": "Internal Links"
          },
          "discovery": {
            "$ref": "#/components/schemas/DiscoveryConfig"
          },
          "schedule": {
            "$ref": "#/components/schemas/ScheduleConfig"
          },
          "approval_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Email",
            "description": "Where draft approval emails go (defaults to account email)."
          }
        },
        "type": "object",
        "required": [
          "name",
          "domain"
        ],
        "title": "SiteIn"
      },
      "SiteOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "brand_voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Voice"
          },
          "custom_prompts": {
            "$ref": "#/components/schemas/CustomPrompts"
          },
          "image_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Style",
            "description": "Visual style guidance for generated images."
          },
          "pillar_cluster_map": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Pillar Cluster Map",
            "description": "pillar -> cluster keywords."
          },
          "internal_links": {
            "items": {
              "$ref": "#/components/schemas/InternalLink"
            },
            "type": "array",
            "title": "Internal Links"
          },
          "discovery": {
            "$ref": "#/components/schemas/DiscoveryConfig"
          },
          "schedule": {
            "$ref": "#/components/schemas/ScheduleConfig"
          },
          "approval_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Email",
            "description": "Where draft approval emails go (defaults to account email)."
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "domain",
          "id"
        ],
        "title": "SiteOut"
      },
      "SiteUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "brand_voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Voice"
          },
          "custom_prompts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomPrompts"
              },
              {
                "type": "null"
              }
            ]
          },
          "image_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Style"
          },
          "pillar_cluster_map": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pillar Cluster Map"
          },
          "internal_links": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/InternalLink"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Links"
          },
          "discovery": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiscoveryConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "approval_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Email"
          }
        },
        "type": "object",
        "title": "SiteUpdate",
        "description": "Partial update \u2014 every field optional."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}